Made thunk action pattern consistent
