+++ src/store/application/applicationThunks.ts
@@ -1,0 +1,9 @@
+import { AppThunk } from "store/rootStore";
+
+import { setPlayerNameAction } from "./applicationActions";
+
+export const setPlayerName =
+  (playerName: string): AppThunk =>
+  (dispatch) => {
+    dispatch(setPlayerNameAction(playerName));
+  };
