+++ src/store/host/hostActions.ts
@@ -1,57 +1,54 @@
 import { createAction } from "@reduxjs/toolkit";
 
 import { GameSettings } from "models/GameSettingsType";
+import { GameStatus } from "models/GameStatus";
 import { Modals } from "models/Modals";
 import { PlayerState } from "models/PlayerState";
-import {
-  PEER_HOST_START,
-  PEER_HOST_STOP,
-} from "store/middleware/middlewareConstants";
+import { SoundEffects } from "utils/SoundEffectList";
 
 import {
   HOST_INCREMENT_ROUND,
-  HOST_FORCED_DISCONNECT,
+  HOST_PEER_FORCED_DC,
   HOST_SELECT_ICON,
   HOST_VISIT_SETTINGS,
   HOST_SUBMIT_PLAYER_NAME,
   HOST_MODIFY_SETTINGS,
+  HOST_PEER_START,
+  HOST_PEER_STOP,
 } from "./hostConstants";
-import {
-  prepareAddSelectedIcon,
-  prepareIncrementRound,
-  prepareVisitSettingsPage,
-} from "./hostHelpers";
 
-export const incrementRound = createAction(
-  HOST_INCREMENT_ROUND,
-  function prepare(overrideRound?: number) {
-    return {
-      payload: prepareIncrementRound(overrideRound),
-    };
-  }
-);
+export const incrementRoundAction = createAction<{
+  round: number;
+  targetHeroes: number[];
+  currentHeroes: number[][];
+  statusText: string;
+  gameStatus: GameStatus;
+  playerName: string;
+}>(HOST_INCREMENT_ROUND);
 
-export const addSelectedIcon = createAction(
-  HOST_SELECT_ICON,
-  function prepare(selectedIcon: number, selectedPlayerName: string) {
-    return {
-      payload: prepareAddSelectedIcon(selectedIcon, selectedPlayerName),
-    };
-  }
-);
+export const addSelectedIconAction = createAction<
+  | {
+      isCorrectHero: boolean;
+      lastClickedPlayerName: string;
+      soundEffect: SoundEffects | undefined;
+      nextRoundTimer: NodeJS.Timer | null;
+      newState: {
+        players: {
+          [x: string]: PlayerState;
+        };
+        selectedIcons: number[];
+        invalidIcons: number[];
+        statusText: string;
+        gameStatus: GameStatus;
+      };
+    }
+  | undefined
+>(HOST_SELECT_ICON);
 
-export const visitSettingsPage = createAction(
-  HOST_VISIT_SETTINGS,
-  function prepare(isClient: boolean, settings: GameSettings) {
-    return {
-      payload: {
-        players: prepareVisitSettingsPage(isClient),
-        isClient,
-        settings,
-      },
-    };
-  }
-);
+export const visitSettingsPageAction = createAction<{
+  players?: Record<string, PlayerState>;
+  settings: GameSettings;
+}>(HOST_VISIT_SETTINGS);
 
 export const modifyGameSettingsAction =
   createAction<GameSettings>(HOST_MODIFY_SETTINGS);
@@ -61,12 +58,12 @@ export const submitPlayerNameAction = createAction<{
   players: Record<string, PlayerState>;
 }>(HOST_SUBMIT_PLAYER_NAME);
 
+export const hostPeerStartAction = createAction<string>(HOST_PEER_START);
+export const hostPeerStopAction = createAction(HOST_PEER_STOP);
+
 // Called when the host is forcefully disconnected
 // Clean up state and display error modal
-export const hostForcefulDisconnect = createAction<{
+export const hostForcefulDisconnectAction = createAction<{
   modal: Modals;
   message?: string[];
-}>(HOST_FORCED_DISCONNECT);
-
-export const startHostWS = createAction<string>(PEER_HOST_START);
-export const stopHostWS = createAction(PEER_HOST_STOP);
+}>(HOST_PEER_FORCED_DC);
