+++ src/store/host/hostConstants.ts
@@ -3,4 +3,6 @@ export const HOST_SELECT_ICON = "host/selectIcon";
 export const HOST_VISIT_SETTINGS = "host/visitGameSettings";
 export const HOST_MODIFY_SETTINGS = "host/modifyGameSettings";
 export const HOST_SUBMIT_PLAYER_NAME = "host/submitPlayerName";
-export const HOST_FORCED_DISCONNECT = "host/forcedPeerDisconnect";
+export const HOST_PEER_START = "host/peerStart";
+export const HOST_PEER_STOP = "host/peerStop";
+export const HOST_PEER_FORCED_DC = "host/peerForcedDisconnect";
