+++ src/store/middleware/middlewareConstants.ts
@@ -1,8 +1,5 @@
 export const PLAY_AUDIO_ACTION = "application/playAudio";
 
-export const PEER_HOST_START = "host/startHosting";
-export const PEER_HOST_STOP = "host/stopHosting";
-
 export const PEER_CLIENT_CONNECT = "client/connect";
 export const PEER_CLIENT_DISCONNECT = "client/disconnect";
 export const PEER_CLIENT_SEND = "client/send";
